requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.3.7
python-dotenv==1.0.0
lxml==4.9.3
readability-lxml==0.8.1
newspaper3k==0.2.8
